Updated RWT to float and add filters to gear-effort-process-type
This commit updates RWT to a float to conform to some data
sources. The commit include the migration file and updates to the
validation schema.  It also include default filters to the
gear-effort-process-type api endpoint to only return active, approved
gears by default, the others are available, but will not be returned
by default. A test has been added to verify the filters work as
expected. All previously passing tests continue to pass.

+++ fn_portal/models/FN1_tables.py
@@ -270,7 +270,7 @@ class FN125(models.Model):
     fish = models.CharField(max_length=6)
     flen = models.IntegerField(blank=True, null=True)
     tlen = models.IntegerField(blank=True, null=True)
-    rwt = models.IntegerField(blank=True, null=True)
+    rwt = models.FloatField(blank=True, null=True)
     girth = models.IntegerField(blank=True, null=True)
     clipa = models.CharField(max_length=20, blank=True, null=True)
     clipc = models.CharField(max_length=20, blank=True, null=True)
